packages/agent/requirements.txt
# AURA AdaptLearn — Python Agent Service Dependencies
# Pinned for reproducible production builds

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
asyncpg==0.29.0

# Cache / messaging
redis==5.0.4

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.2.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.3

# AI provider
anthropic==0.26.0

# Object storage
minio==7.2.7

# HTTP client
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
